requirements.txt
# ─── Core ML ──────────────────────────────────────────────────────────────────
tensorflow<2.11.0
numpy<1.24.0
protobuf==3.20.3
scikit-learn>=1.0.0
Pillow>=10.0.0

# ─── Explainability ───────────────────────────────────────────────────────────
opencv-python-headless>=4.8.0

# ─── Backend API ──────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0

# ─── AI Integration ───────────────────────────────────────────────────────────
groq>=0.9.0
tavily-python>=0.3.0

# ─── Frontend ─────────────────────────────────────────────────────────────────
streamlit>=1.35.0
plotly>=5.22.0

# ─── Data Engineering (Dipertahankan untuk Notebooks) ─────────────────────────
pandas>=2.1.0
seaborn>=0.13.0
matplotlib>=3.8.0

# ─── Environment & Config ─────────────────────────────────────────────────────
python-dotenv>=1.0.0